+++ packages/client/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Removed
+
+- exported `TxCall` interface, inlined directly as `(): Promise<Uint8Array>`
+
 ## 2.0.0-rc.3 - 2026-03-04
 
 ### Fixed