+++ packages/client/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2.0.0-rc.4 - 2026-03-11
+
 ### Changed
 
 - Replace replace `decodeCallData` for `txFromCallData` in tx staticApis.